+++ loqusdb/commands/update.py
@@ -62,6 +62,13 @@ LOG = logging.getLogger(__name__)
     show_default=True,
     help="Add SNVs to an existing case without replacing its stored SNV VCF",
 )
+@click.option(
+    "--ignore-gq-if-unset",
+    is_flag=True,
+    default=False,
+    show_default=True,
+    help="Ignore GQ threshold if GQ is unset in VCF",
+)
 @click.pass_context
 def update(
     ctx,
@@ -75,6 +82,7 @@ def update(
     ensure_index,
     max_window,
     add_to_existing_snv,
+    ignore_gq_if_unset,
 ):
     """Load the variants of a case
 
@@ -121,6 +129,7 @@ def update(
             genome_build=genome_build,
             keep_chr_prefix=keep_chr_prefix,
             add_to_existing_snv=add_to_existing_snv,
+            ignore_gq_if_unset=ignore_gq_if_unset,
         )
     except (SyntaxError, CaseError, IOError, VcfError) as error:
         LOG.warning(error)